Wait for all tasks to finish on shutdown
Previously, we've waited only for explicit requests, but not for the
notifications. That is, it was possible to receive "file changed"
notification, schedule the build, receive "shutdown" request, respond
to shutdown, and *then* send a bunch of diagnostics from the compiler.

This commit simplifies this logic so that we wait for *all* things
before processing a blocking notification.